Add proper Marshal, Unmarshal and (De|En)coders
Verify now uses the new Unmarshal function.

Marshaling to and unmarshalling from string was easy to get started and
focus on implementing the validation, but it's time to introduce proper
API.

SecureLogin tokens are supposed to be passed over network and in Go this
happens by reading and writing bytes. Underneath still plain casting to
string happens but since sltokens are expected to be quite small I don't
think this would be an actual issue.

Also, add examples how to use encoders and decoders.